feat: use L2EthTransactionMapper inherit EthTransactionMapper to make code more reasonable
