+++ ethereumetl/domain/transaction.py
@@ -37,3 +37,10 @@ class EthTransaction(object):
         self.max_fee_per_gas = None
         self.max_priority_fee_per_gas = None
         self.transaction_type = None
+
+class L2EthTransaction(EthTransaction):
+    def __int__(self):
+        super().__int__()
+        self.l1_block_number = None
+        self.l1_tx_origin = None
+        self.l1_times_stamp = None
